requirements.txt
langchain-openai
langchainhub
langchain-chroma
langchain
playwright
beautifulsoup4
ipykernel